Generate test cases on build reproducibly.
This change will allow avoiding committing binary blobs to the repo
by ensuring that we use a well defined gcc (via docker) version for
creation of binaries from c source snippets.

The Makefile has been updated to run a sha256 check to ensure
that the built blobs are exactly as expected. When adding tests, the
.expected_shasums file will need to be updated.