Add an option to filter on label names
This change adds a `source-filter` option which allows a regex to be
provided to filter the names of keys in data sources. When provided,
only keys matching the pattern will be counted as part of the reported
size. The option can be set via config files or the command line.