tests: handle another truncation warning
MSVC flags this truncation as the size of the file is tracked as a
`uint64_t`, which is then used to construct a tuple where the size
is used in a field marked as `int`, resulting in a truncation if a
large file is encountered (which we know to not be the case here).
This statically casts away the truncation to silence it.