correct an assertion for ILP32 targets
The assertion was incorrectly written as `N <= sizeof(N)`.  When
building for 64-bit platforms, the `N` would implicitly be 64-bits (the
default width of an integer).  However, on a 32-bit build, `sizeof(N)`
would be 4, with `N` however being 8 in the case of a 64-bit read.  This
assertion is meant to ensure that the data being read fits into the
destination type and should be `N <= sizeof(T)`.  This repairs the
32-bit builds of bloaty.