CI: adjust the testing command
This adjusts the testing command to match the existing Travis CI
invocation of "make test".  Use `cmake` to drive the test target
to allow using a different make program more easily.  On Windows
we need to special case the target name unfortunately as MSBuild
fails to execute the tests otherwise.