tests: repair `fuzz_driver` execution on Windows
The fuzz_driver used `std::ifstream` with the single parameter
constructor.  This defaults the second parameter to `std::ios_base::in`
which will process the file in text mode.  This works fine for most
platforms as they largely process input as UTF-8.  However, Windows
defaults to UTF-16 and will fail to process the input as a byte stream
resulting in a failure to load the corpus.  Explicitly set the input
mode to `std::ios_base::in | std::ios_base::binary` to ensure that we
can read the input on all the platforms.  This repairs the fuzzing tests
on Windows.