CI: add a workaround for macOS builders on GHA
The GHA agents have installed capstone via brew.  Unfortunately, the
pkg-config data for the bottle provides an include path with an extra
component which we cannot reliably strip in the build system.  The
result is that `capstone/capstone.h` is not found.  Explicitly inject an
additional header search path from the command line to resolve the
header properly.