Fix infinite loop in ReadEhFrameHdr when table_enc is omit
When table_enc is DW_EH_PE_omit (0xff) in .eh_frame_hdr, ReadEncodedPointer
returns 0 without advancing the data pointer. If fde_count is large (e.g.
from a fuzzed binary), this leads to an infinite loop in ReadEhFrameHdr as
it attempts to read fde_count entries.

This change adds a check to return early from ReadEhFrameHdr if table_enc
is DW_EH_PE_omit, as there is no table to read in this case.

Fixes #453.