Allow undefined symbol names
When processing symbol table names an error would be thrown if the the
name was undefined (0) resulting in failures to process files like:

> bloaty: can't read index 0 from strtab, total size is 254049

This modifies the symbol table processing to return an empty string if
the name is undefined. An empty string will trigger the existing logic
to generate a synthetic "Anonymous" name instead of crashing.

Fixes: #217
Fixes: #345